:recycle: 전역 url을 활용하여 상세 페이지 게스트 접근 가능하도록 컴포넌트 구조 리팩터링
